A part of testing for intersection algorithm implemented
transform_triangle() function belongs to anonymous namespace. The opening and the closure of this namespace were surrounded by
preprocessor directives to make testing any function in there possible. Moreover, a specific header was implemented to contain
prototypes of functions from anonymous namespace to be tested. Tests for transform_triangle() were implemented. Because of all
these novelties, the sructure of test-directory was changed.